DLCs_Main: Split Christmas98 and Christmas99 into separate files.
DLCs.cpp still needs matlist_00000004_2 from Data/DLC_Christmas99.h,
so it's now declared as `extern`.
