+++ DLCs_Main/Christmas98.h
@@ -1,0 +1,8 @@
+// Christmas '98 DLC functions.
+#pragma once
+
+#include <SADXStructs.h>
+extern const PVMEntry Christmas98Textures;
+
+void Christmas98_Load(ObjectMaster *a1);
+void Christmas98_LoadStuff(void);
