+++ DLCs_Main/Christmas99.h
@@ -1,0 +1,8 @@
+// Christmas '98 DLC functions.
+#pragma once
+
+#include <SADXStructs.h>
+extern const PVMEntry Christmas99Textures;
+
+void Christmas99_Load(ObjectMaster *a1);
+void Christmas99_LoadStuff(void);
