+++ DLCs_Main/DLCs_Main.vcxproj
@@ -76,6 +76,8 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="ATT.cpp" />
+    <ClCompile Include="Christmas98.cpp" />
+    <ClCompile Include="Christmas99.cpp" />
     <ClCompile Include="Balloon.cpp" />
     <ClCompile Include="DLCs.cpp" />
     <ClCompile Include="Messages.cpp" />
@@ -89,6 +91,8 @@
   <ItemGroup>
     <ClInclude Include="ATT.h" />
     <ClInclude Include="Balloon.h" />
+    <ClInclude Include="Christmas98.h" />
+    <ClInclude Include="Christmas99.h" />
     <ClInclude Include="Data\DLC_ATT.h" />
     <ClInclude Include="Data\DLC_Christmas98.h" />
     <ClInclude Include="Data\DLC_Christmas99.h" />
